add json_response method call
When a user would send a POST request to create a new todo item, the
Create method would not send anything back. Because of this, the spec
'returns an http response (200)' would fail as the http response code
would be (204) meaning content was expected but none was returned. This
commit returns the todo item created by the user.