move get requests into context
The reason I did this is because I want to test the edge cases of
getting a list of Todos. Such as when there isn't a todo object in the
database or (eventually) the todo is not owned by you. I do not think
this is standard coding practice but HEY! If it works, don't fix it.

I'm 100% certain a professional would be very angry at this commit.