edit test to expect 404 and pass
I changed the Spec and test to expect a different http response status
because I originally didn't understand how Rails handles trying to find
an object in the database that doesn't exist. I was going to make my own
validation method and return the http response status (400) but it is
not needed because Rails returns (404) regardless.